CP-54351 move nfs from /opt
