+++ Makefile
@@ -48,6 +48,7 @@ SM_LIBS += lvmcache
 SM_LIBS += lvutil
 SM_LIBS += metadata
 SM_LIBS += mpathcount
+SM_LIBS += nfs
 SM_LIBS += pluginutil
 SM_LIBS += refcounter
 SM_LIBS += resetvdis
